requirements.txt
# Web scraping and HTTP
requests>=2.28.0
beautifulsoup4>=4.11.0

# Image processing
Pillow>=9.2.0
opencv-python>=4.6.0
imagehash>=4.3.0

# Data processing and analysis
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.1.0

# Data storage
pyarrow>=9.0.0

# Visualization
matplotlib>=3.5.0

# Utility
tqdm>=4.64.0
python-dotenv>=0.21.0